+++ js/recurring.js
@@ -46,6 +46,15 @@ Recurring Payments field
 		$("#ui-datepicker-div").hide();
 	}
 
+	/**
+	* set recurring field to active, until otherwise advised by conditional logic triggers
+	* @param {jQuery.Event} event
+	* @param {Number} form_id int ID of Gravity Forms form
+	*/
+	$(document).on("gform_post_render", function(event, form_id) {
+		$("#gform_" + form_id + " .gfeway-contains-recurring").addClass("gfeway-recurring-active");
+	});
+
 	// watch for conditional logic changes
 	gform.addAction("gform_post_conditional_logic_field_action", function(formId, action, targetId, defaultValues, isInit) {
 		var target = $(targetId);